requirements.txt
PyQt5==5.15.9
pycryptodome==3.19.0